requirements.txt
# ============================================================
# Bone Fracture Classification - Requirements
# KBG Hackathon 2025 | IIT Mandi
# Pinned to exact versions used in the current environment
# ============================================================
torch==2.9.1
torchvision==0.24.1
timm==1.0.25
numpy==2.4.2
pandas==2.3.3
scikit-learn==1.7.2
matplotlib==3.10.7
seaborn==0.13.2
Pillow==12.1.1
PyYAML==6.0.3
tqdm==4.67.1
albumentations==2.0.8
ImageHash>=4.3.1
opencv-python-headless==4.13.0.92
scipy==1.16.3
streamlit==1.55.0
